ops-incident-major-analysis/requirements.txt
# Major Incident Analysis Report - Dependencies

# Core
python-dotenv>=1.0.0

# Data processing
# (minimal dependencies - JSON is built-in)

# Report generation
jinja2>=3.1.0
python-docx>=0.8.11

# AI insights (optional)
openai>=1.0.0

# Development
pytest>=7.0.0
